==> src/04_modelviz.qmd <==
---
title: "3. Model"
format: html
---

Import packages:

```{r}
library(here)
library(tidyr)
library(dplyr)
library(rlang)
library(iml)
library(glue)
library(brms)
library(purrr)
library(stringr)
library(marginaleffects)
library(ggplot2)
library(readr)
library(tibble)
source(here("src", "utilities.R"))
```

Import data:

```{r}
data = readRDS(here("data", "data_processed.rds"))
pron_order = c("Jun", "Qing", "Gong", "RuEr")
data = data %>% mutate(PronTypeSimple = factor(PronTypeSimple, levels = pron_order))
```

```{r}
#model_pseudobma_weights = readRDS(here("models", "model_best3_pseudobma_weights.rds"))
load_models(regex = "^model_main_powkin\\.rds$")
```

# Effects

```{r}
getFixedEffects = function(model){
    data.frame(fixef(model, probs = c(.1, .9))) %>%
        tibble::rownames_to_column()
}
getRandEffects = function(model){
    data.frame(coef(model, probs = c(.1, .9))) %>%
        tibble::rownames_to_column()
}

write_csv(getFixedEffects(model_main_powkin), here("models", "model_powkin_fe.csv"))
write_csv(getRandEffects(model_main_powkin), here("models", "model_powkin_re.csv"))
```

# Compute predictions and comparisons

```{r}
data_x = data %>% select(Speaker, Power, Intimate, StanceH, StanceS, BenH, Kin)
```

```{r}
predictor_combos = tidyr::expand_grid(Power = c("Speaker", "Equal", "Listener"), Intimate = c("/", "Y", "N"), Kin = c("Yes", ""),
    StanceH = c("Pos", "Neg", "Q", ""), StanceS = c("Pos", "Q", "")) %>%
        filter(!(Kin ==  "" & Intimate == "Y"))
datagrid_allcombosexceptimpose_speakerunk = datagrid(Power = predictor_combos$Power, Intimate = predictor_combos$Intimate, Kin = predictor_combos$Kin,
    StanceH = predictor_combos$StanceH, StanceS = predictor_combos$StanceS, Speaker = "?", model = model_main_powkin)
write_csv(datagrid_allcombosexceptimpose_speakerunk, here("output", "datagrid_allcombosexceptimpose_speakerunk.csv"))
```

```{r}
predicts_allcombosexceptimpose_speakerunk = predictions(model_main_powkin, datagrid_allcombosexceptimpose_speakerunk,
        conf.level = .8)
write_csv(predicts_allcombosexceptimpose_speakerunk, here("output", "predicts_allcombosexceptimpose_speakerunk.csv"))
```

```{r}
predicts_allcombosexceptimpose_speakerunk_50 = predictions(model_main_powkin, datagrid_allcombosexceptimpose_speakerunk,
        conf.level = .5)
write_csv(predicts_allcombosexceptimpose_speakerunk_50, here("output", "predicts_allcombosexceptimpose_speakerunk_50.csv"))
```

```{r}
predicts_orig = predictions(model_main_powkin, conf.level = .8)
write_excel_csv(predicts_orig, here("output", "predicts_orig.csv"))
```

```{r}
comparisons_all = comparisons(model_main_powkin,
                                newdata = datagrid_allcombosexceptimpose_speakerunk,
                                variables = list(Power = "all", Intimate = "all",
                                                 StanceS= "all", StanceH = "all"),
                                                 conf_level = 0.8)
write_csv(comparisons_all, here("output", "model_powkin_comparisons_all.csv"))
```

```{r}
comparisons_all_50 = comparisons(model_main_powkin,
                                newdata = datagrid_allcombosexceptimpose_speakerunk,
                                variables = list(Power = "all", Intimate = "all",
                                                 StanceS= "all", StanceH = "all"),
                                                 conf_level = 0.5)
write_csv(comparisons_all_50, here("output", "model_powkin_comparisons_all_50.csv"))
```

# Visualisation

```{r}
comparisons_all = read_csv(here("output", "model_powkin_comparisons_all.csv"))
predicts = read_csv(here("output", "predicts_allcombosexceptimpose_speakerunk.csv"))
comparisons_all_50 = read_csv(here("output", "model_powkin_comparisons_all_50.csv"))
predicts_50 = read_csv(here("output", "predicts_allcombosexceptimpose_speakerunk_50.csv"))
predicts_orig = read_csv(here("output", "predicts_orig.csv"))
```

```{r}
combine_conflevels = function(df_80, df_50){
    df_80 %>%
        cbind(df_50 %>%
             rename(conf.low_50 = conf.low, conf.high_50 = conf.high) %>%
             select(conf.low_50, conf.high_50))
 }
comparisons_all = combine_conflevels(comparisons_all, comparisons_all_50)
predicts = combine_conflevels(predicts, predicts_50)
```

```{r}
prep_me_df_for_viz = function(df){
    df %>%
        mutate(Kin = replace_na(Kin, "No")) %>%
        mutate(RelDist = case_when(Kin == "Yes" ~ "Kin",
                               Intimate == "Y" & Kin == "No" ~ "IntimateNonKin",
                               Intimate == "N" & Kin == "No" ~ "NonIntimate")) %>%
        mutate(group = factor(group, levels = c("Qing", "RuEr", "Jun", "Gong"))) %>%
        mutate(RelDist = factor(RelDist, levels = c("Kin", "IntimateNonKin", "NonIntimate")),
               Power = factor(Power, levels = c("Listener", "Equal", "Speaker"))) %>%
        filter(!is.na(RelDist)) %>%
        filter((!(Kin == "Yes" & Intimate == "N") & !(Kin == "Yes" & Intimate == "/")) | is.na(Kin))
}
comparisons_viz = prep_me_df_for_viz(comparisons_all)
predicts_viz = prep_me_df_for_viz(predicts)
```

```{r}
levels2text_RelDist = c(Kin = "Kin", IntimateNonKin = "Intimate Non-Kin", NonIntimate = "Non-Intimate")
levels2text_pow = c(Listener = "Listener", Speaker = "Speaker", Equal = "Equal")
levels2text_pron = c(Qing = "qīng", RuEr = "rǔ/\něr", Jun = "jūn", Gong = "gōng")
```

Predictions for each dyad type with neutral stance:

```{r}
predicts_viz %>%
    filter(is.na(StanceH), is.na(StanceS), is.na(ImpositionH)) %>%
    ggplot(aes(x = group, y = estimate)) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2, linewidth = .3) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 1) +
    theme_bw() + 
    #geom_line(aes(group = Power)) +
    ylab("Probability") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron)
ggsave(here("output", "predict_by_dyad.svg"), width = 868L, height = 546L, units = "px")
```

I want to compare StanceS values in different dyads:

```{r}
predicts_viz %>%
    filter(StanceH %in% c("Pos", "Neg"), is.na(StanceS), is.na(ImpositionH)) %>%    ggplot(aes(x = group, y = estimate, group = StanceH, col = StanceH)) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point() +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2) +
    theme_bw() + 
    geom_line() +
    ylab("Probability") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron)
ggsave(here("output", "predict_stanceh_by_dyad.svg"), width = 868L, height = 546L, units = "px")

```

```{r}
comparisons_viz %>%
    filter(term == "StanceH", contrast %in% c("Neg -", "Pos -"), is.na(StanceS), is.na(ImpositionH)) %>%
    ggplot(aes(x = group, y = estimate, group = contrast, col = contrast)) +
    geom_hline(yintercept = 0) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2, position = position_dodge(.9)) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2, alpha = .8, position = position_dodge(.9)) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 1.5, alpha = .8, position = position_dodge(.9)) +
    theme_bw() + 
    #geom_line() +
    ylab("Probability difference\n(vs. neutral stance)") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron) +
    scale_color_discrete(name = "Stance", labels = c("Neg", "Pos"))
ggsave(here("output", "compare_stanceh_by_dyad.svg"), width = 868L, height = 546L, units = "px")

```

```{r}
comparisons_viz %>%
    filter(term == "StanceS", contrast %in% c("Pos -"), is.na(StanceH), is.na(ImpositionH)) %>%
    ggplot(aes(x = group, y = estimate, group = contrast)) +
    geom_hline(yintercept = 0) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2, alpha = .8, linewidth = .3) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 1.5, alpha = .8) +
    theme_bw() + 
    geom_line() +
    ylab("Probability difference\n(negative vs. neutral stance)") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron)
ggsave(here("output", "compare_stanceh_negvsnone_by_dyad.svg"), width = 868L, height = 546L, units = "px")
```

```{r}
comparisons_viz %>%
    filter(term == "StanceH", contrast %in% c("Pos -"), is.na(StanceS)) %>%
    ggplot(aes(x = group, y = estimate, group = contrast)) +
    geom_hline(yintercept = 0) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2, alpha = .8, linewidth = .3) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 1.5, alpha = .8) +
    theme_bw() + 
    geom_line() +
    ylab("Probability difference\n(positive vs. neutral stance)") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron)
ggsave(here("output", "compare_stanceh_posvsnone_by_dyad.svg"), width = 868L, height = 546L, units = "px")

```

```{r}
comparisons_viz %>%
    filter(term == "StanceS", contrast %in% c("Pos -"), is.na(StanceH)) %>%
    ggplot(aes(x = group, y = estimate)) +
    geom_hline(yintercept = 0) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 2) +
    theme_bw() + 
    geom_line() +
    ylab("Probability difference\n(positive vs. neutral stance)") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron) +
    theme()
ggsave(here("output", "compare_stances_by_dyad.svg"), width = 868L, height = 546L, units = "px")

```


```{r}
comparisons_viz %>%
    filter(term == "ImpositionT", contrast %in% c("T -"), is.na(StanceH), is.na(StanceS)) %>%
    ggplot(aes(x = group, y = estimate, col = contrast)) +
    geom_hline(yintercept = 0) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 2) +
    theme_bw() + 
    geom_line() +
    ylab("Probability difference\n(compared to neutral stance)") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron) +
    theme()
ggsave(here("output", "compare_imposeT_by_dyad.svg"), width = 868L, height = 546L, units = "px")

```


```{r}
comparisons_viz %>%
    filter(term == "StanceS", contrast %in% c("HBen -"), is.na(StanceH), is.na(StanceS)) %>%
    ggplot(aes(x = group, y = estimate, col = contrast)) +
    geom_hline(yintercept = 0) +
    facet_grid(Power ~ RelDist,
               labeller = as_labeller(c(levels2text_RelDist, levels2text_pow))) +
    geom_point(size = 2.2) +
    geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = .2) +
    geom_errorbar(aes(ymin = conf.low_50, ymax = conf.high_50), width = 0, linewidth = 2) +
    theme_bw() + 
    geom_line() +
    ylab("Probability difference\n(compared to neutral stance)") +
    theme(text = element_text(size = 26)) +
    scale_x_discrete(labels = levels2text_pron) +
    theme()
ggsave(here("output", "compare_imposeHBen_by_dyad.svg"), width = 868L, height = 546L, units = "px")

```

# Local interpretation

Generate predictor object for `iml`:
```{r}
X = data %>%
    select(Power, Intimate, StanceH, StanceS, Kin, ImpositionH, Speaker)
predictor = Predictor$new(model_main_powkin, data = X, y = data$PronTypeSimple)

```

Generate Shapley values:
```{r}
shaplist = list()
examples_to_use = c(71, 104, 12, 84, 340, 331, 113, 124, 93, 234, 344, 122, 97, 224, 131, 112, 124, 133, 237, 344, 122, 97, 224, 88, 61, 43, 127, 335, 73, 261, 72, 142, 90, 215, 163, 89)
for(x in examples_to_use){
    print(glue("Now calculating Shapley value for data no.: {x}"))
    shaplist[[x]] = Shapley$new(predictor, x.interest = X[x, ], sample.size = 50)
}
shapley_values = saveRDS(shaplist, here("output", "shaplist.rds"))
```

```{r}
shaplist = readRDS(here("output", "shaplist.rds"))
```

```{r}
varnameOrder = c("Power", "Intimate", "Kin", "StanceH", "StanceS", "ImpositionH", "Speaker")
getWaterfall = function(shap, varnameMap = NULL, varnameOrder = NULL, varnameReplace = NULL){
  graph = shap$results %>%
      filter(class == "P(Y = Qing)") %>%
      mutate(feature = factor(feature, levels = c("Power", "Intimate", "Kin", "StanceH", "StanceS", "ImpositionH", "Speaker"))) %>%
      arrange(feature) %>%
      select(-class, -feature, -phi.var) %>%
      mutate(phi = round(phi, 3)) %>%
      waterfall(linetype = "dotted", rect_border = "grey", rect_text_size = 2) +
      theme_minimal() + theme(text=element_text(size=21)) + scale_y_continuous(labels = \(x) round(x + 0.472382, 2), "Accumulated prediction") +
      xlab("predictor")

    #   scale_y_continuous(limits = c(cs_min, cs_max)) +
    #   
    if(!is.null(varnameReplace))
        graph = graph  +
            scale_x_discrete(labels = varnameReplace)

    graph
}
```

```{r}
varnameReplace_84 = c("Power=\nEqual", "Intimate=\nY",  "Kin=\nNo","StanceH=\nNeg", "StanceS=\nN/A", "ImpositionH=\nN/A", "Speaker=\nKǒng Chúnzhī")
getWaterfall(shap, varnameOrder = varnameOrder, varnameReplace = varnameReplace_84)
```

```{r}
varnameReplace_61 = c("Power=\nListener", "Intimate=\nN", "Kin=\nN", "StanceH=\nNeg", "StanceS=\nN/A", "ImpositionH=\nN/A", "Speaker=\nLù Jī")
getWaterfall(shaplist[[61]], varnameOrder = varnameOrder, varnameReplace = varnameReplace_61)
```


```{r}
varnameReplace_104 = c("Power=\nListener", "Intimate=\nN", "Kin=\nN", "StanceH=\nNeg", "StanceS=\nN/A", "ImpositionH=\nN/A", "Speaker=\nLù Jī")
getWaterfall(shaplist[[104]], varnameOrder = varnameOrder)
##NOTE: Fix this in annos!!!!!
```
